Tailor build script to fsl-strace
Previously, this repository `fsl-strace` had a build-script named
`build-syscall-replayer` located in the `dist` folder that installed all
dependencies required to build Re-Animator's syscall-replayer. There are
a number of idiosyncrasies with this method.
1. An outdated `build-fsl-strace` script remained in the repository
root, while the correct build script was hidden away in `dist/`.
2. The `fsl-strace` repository was downloaded as part of the build
process, even though the repository had to have been downloaded to get
access to the script.
3. The build instructions must make reference to syscall-replayer.
Although that is the whole purpose of the Re-Animator project, it is not
required for the fsl-strace binary to function.

This commit changes the build script to be specific to fsl-strace. The
old version of the build script will be migrated to syscall-replayer in
the trace2model repository.